Switch GarNet to new API
